feat: 불필요한 코드 제거 및 콘솔 제거
